move random stuff to RandomGraph component
