+++ frontend/src/components/RandomGraph.tsx
@@ -1,19 +1,33 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useVisNetwork } from '../hooks/useVisNetwork'
 
 const Graph: React.FC = () => {
-  const { containerRef, nodesRef, edgesRef } = useVisNetwork({
+  const { containerRef, networkRef, nodesRef, edgesRef } = useVisNetwork({
     nodes: [],
     edges: [],
   })
 
-  const handleAddNode = () => {
+  const addRandomNode = () => {
+    console.log(`handleAddNode`, nodesRef)
     const numNodes = nodesRef.current.length
     const newNodeId = numNodes + 1
     nodesRef.current.add({ id: newNodeId, label: `Node ${newNodeId}` })
+
+    if (numNodes === 0) {
+      networkRef.current?.fit()
+    } else {
+      const nodeIds = nodesRef.current.getIds()
+      const fromNodeId = nodeIds[Math.floor(Math.random() * nodeIds.length)]
+
+      edgesRef.current.add({
+        id: `${fromNodeId} -> ${newNodeId}`,
+        from: fromNodeId,
+        to: newNodeId,
+      })
+    }
   }
 
-  const handleAddEdge = () => {
+  const addRandomEdge = () => {
     const numEdges = edgesRef.current.length
     const newEdgeId = numEdges + 1
 
@@ -25,12 +39,30 @@ const Graph: React.FC = () => {
     edgesRef.current.add({ id: newEdgeId, from: fromNodeId, to: toNodeId })
   }
 
+  useEffect(() => {
+    const intervalId = setInterval(() => {
+      if (Math.random() < 0.2) {
+        addRandomEdge()
+      } else {
+        addRandomNode()
+      }
+    }, 500)
+
+    setTimeout(() => {
+      clearInterval(intervalId)
+    }, 20_000)
+
+    return () => {
+      clearInterval(intervalId)
+    }
+  }, [])
+
   return (
     <div>
       <div ref={containerRef} style={{ width: '800px', height: '600px' }} />
       <div>
-        <button onClick={handleAddNode}>Add Node</button>
-        <button onClick={handleAddEdge}>Add Edge</button>
+        <button onClick={addRandomNode}>Add Node</button>
+        <button onClick={addRandomEdge}>Add Edge</button>
       </div>
     </div>
   )
