+++ frontend/src/components/WebsocketGraph.tsx
@@ -14,16 +14,31 @@ const Graph: React.FC = () => {
 
   const onMessage = useCallback<NonNullable<WebSocketHandlers['onMessage']>>(
     (event) => {
-      const { wikid }: PageData = JSON.parse(event.data)
+      const { wikid, children }: PageData = JSON.parse(event.data)
       console.log('onMessage', { wikid, nodesRef, networkRef })
       try {
         nodesRef.current.add({ id: wikid, label: wikid })
-        networkRef.current?.fit()
       } catch (err) {
         console.warn(err)
       }
-      networkRef.current?.fit()
-      networkRef.current?.redraw()
+      for (const child of children) {
+        try {
+          nodesRef.current.add({ id: child, label: child })
+        } catch (err) {
+          console.warn(err)
+        }
+        try {
+          edgesRef.current.add({
+            from: wikid,
+            to: child,
+            id: `${wikid} -> ${child}`,
+          })
+        } catch (err) {
+          console.warn(err)
+        }
+      }
+      // networkRef.current?.fit()
+      // networkRef.current?.redraw()
     },
     [networkRef, nodesRef],
   )
@@ -34,7 +49,7 @@ const Graph: React.FC = () => {
     }
   }, [onMessage])
 
-  const { messages, send } = useWebSocket({
+  const { send } = useWebSocket({
     url: 'wss://flvn62nuq8.execute-api.us-east-1.amazonaws.com/dev',
     handlers,
   })
@@ -60,63 +75,9 @@ const Graph: React.FC = () => {
     setInputValue('')
   }
 
-  const addRandomNode = () => {
-    console.log(`handleAddNode`, nodesRef)
-    const numNodes = nodesRef.current.length
-    const newNodeId = numNodes + 1
-    nodesRef.current.add({ id: newNodeId, label: `Node ${newNodeId}` })
-
-    if (numNodes === 0) {
-      networkRef.current?.fit()
-    } else {
-      const nodeIds = nodesRef.current.getIds()
-      const fromNodeId = nodeIds[Math.floor(Math.random() * nodeIds.length)]
-
-      edgesRef.current.add({
-        id: `${fromNodeId} -> ${newNodeId}`,
-        from: fromNodeId,
-        to: newNodeId,
-      })
-    }
-  }
-
-  const addRandomEdge = () => {
-    const numEdges = edgesRef.current.length
-    const newEdgeId = numEdges + 1
-
-    // Choose two random nodes to connect with a new edge
-    const nodeIds = nodesRef.current.getIds()
-    const fromNodeId = nodeIds[Math.floor(Math.random() * nodeIds.length)]
-    const toNodeId = nodeIds[Math.floor(Math.random() * nodeIds.length)]
-
-    edgesRef.current.add({ id: newEdgeId, from: fromNodeId, to: toNodeId })
-  }
-
-  // useEffect(() => {
-  //   const intervalId = setInterval(() => {
-  //     if (Math.random() < 0.2) {
-  //       addRandomEdge()
-  //     } else {
-  //       addRandomNode()
-  //     }
-  //   }, 500)
-
-  //   setTimeout(() => {
-  //     clearInterval(intervalId)
-  //   }, 20_000)
-
-  //   return () => {
-  //     clearInterval(intervalId)
-  //   }
-  // }, [])
-
   return (
     <div>
       <div ref={containerRef} style={{ width: '800px', height: '600px' }} />
-      <div>
-        <button onClick={addRandomNode}>Add Node</button>
-        <button onClick={addRandomEdge}>Add Edge</button>
-      </div>
       <form onSubmit={handleSubmit}>
         <input
           type="text"
@@ -125,11 +86,11 @@ const Graph: React.FC = () => {
         />
         <button type="submit">Send</button>
       </form>
-      <ul>
+      {/* <ul>
         {messages.reverse().map((message, index) => (
           <li key={index}>{message}</li>
         ))}
-      </ul>
+      </ul> */}
     </div>
   )
 }
